+++ src/main/java/com/depromeet/booboo/ui/controller/MemberController.java
@@ -1,5 +1,8 @@
-package com.depromeet.booboo.member;
+package com.depromeet.booboo.ui.controller;
 
+import com.depromeet.booboo.application.MemberService;
+import com.depromeet.booboo.domain.member.Member;
+import com.depromeet.booboo.ui.dto.common.ApiResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
